% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot_functions.R
\name{add_segments}
\alias{add_segments}
\title{Add additional line segments}
\usage{
add_segments(output, T_x = mean, out = F, ...)
}
\arguments{
\item{output}{the list output from \code{cdf_curve},
\code{pdf_curve}, or \code{hazard_curve}.}

\item{T_x}{a function to calculate a text statistic over
a vector (e.g. mean, median, etc.).}

\item{out}{A logical value, indicating if output should be returned.}

\item{...}{additional plotting parameters.}
}
\value{
A list consisting of the x and y-axis plotting points.
}
\description{
Adds additional vertical line segments to an existing plot,
based on a given test statistic for response times conditioned
on choice using output from a curve function.
}
\examples{
# Load in example dataset
data("priming_data")
d = priming_data
layout( cbind(1,2) )
# Single subject
sel = d$Condition == 15 & d$Subject == 1
rt = d$RT[sel]; ch = d$Accuracy[sel]
blankRTplot()
out = cdf_curve( rt, ch, opt = list( out = T ) )
add_segments( out, col = 'blue' )
add_segments( out, T_x = median, col = 'red' )
legend( 'topleft', c('Mean','Median'), fill = c('blue','red'), bty = 'n' )
# Aggregating over multiple subjects
sel = d$Condition == 15
rt = d$RT[sel]; ch = d$Choice[sel]; grp = d$Subject[sel]
blankRTplot()
out = cdf_curve( rt, ch, opt = list( out = T ) )
T_x = function(x) quantile(x,prob=seq(.1,.9,.1))
add_points( out, T_x = T_x, col = 'grey' )
}

