% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{add_to_rtplot}
\alias{add_to_rtplot}
\title{Add Points or Lines to a RT Plot}
\usage{
add_to_rtplot(object, T_x = mean, T_g = mean, val = NULL, ind = 1,
  type = "p", out = T, ...)
}
\arguments{
\item{object}{a rtplots object or a list with the
x and y-axis values to plot.}

\item{T_x}{the test statistic(s) to be added to the
existing figure.}

\item{T_g}{the test statistic to use when collapsing
over a grouping factor.}

\item{val}{Forthcoming.}

\item{ind}{Forthcoming.}

\item{type}{Forthcoming.}

\item{out}{Forthcoming.}

\item{...}{Forthcoming.}
}
\description{
A convenience function that draws points or
line segments onto an already existing plot. Useful
for adding summary statistics (e.g., the mean or
median).
}
