% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot_functions.R
\name{add_uncertainty}
\alias{add_uncertainty}
\title{Add uncertainty intervals to plot}
\usage{
add_uncertainty(output, f = NULL, alpha = 0.95, opt = list(), ...)
}
\arguments{
\item{output}{the list output from functions like \code{cdf_curve}
or \code{quantile_points}.}

\item{f}{an optional function to calculate the lower and upper boundaries
of an uncertainty interval given a coverage interval.}

\item{alpha}{the desired coverage interval for the uncertainty interval.}

\item{opt}{a list of named options:
\describe{
  \item{\code{draw}}{If true, the intervals are drawn.}
  \item{\code{out}}{If true, output is returned.}
}}

\item{...}{additional plotting parameters.}
}
\value{
A list consisting of...
\describe{
  \item{\code{pv}}{a data frame with the plotting values
    used for the x-axis and the y-axis.}
  \item{\code{ui}}{a list giving the uncertainty intervals}
  \item{\code{opt}}{a list of the options used.}
  }
}
\description{
Adds uncertainty intervals (e.g. the mean +/- 1.96 standard errors)
to a plot using output from other functions that were applied with
a grouping factor.
}
\examples{
# Load in example dataset
data("priming_data")
d = priming_data
layout( cbind(1,2) )
# Select condition
cnd = d$Condition == 3
rt = d$RT[cnd]; ch = d$Accuracy[cnd]; grp = d$Subject[cnd]
# For curve functions
blankRTplot()
out = cdf_curve( rt, ch, grp = grp, opt = list( out = T, draw = F ) )
add_uncertainty( out, border = NA, col = 'grey' )
cdf_curve( rt, ch, grp = grp ) # Now draw the curve
# For point functions (with new type of uncertainty interval)
f = function(x,alpha) {
  prb = c( (1-alpha)/2,alpha+(1-alpha)/2 );
  quantile(x,prb)
}
blankRTplot()
out = quantile_points( rt, ch, grp = grp, opt = list( draw = F, out = T ) )
add_uncertainty( out, f = f, alpha = .5 )
quantile_points( rt, ch, grp = grp, pch = 21, bg = 'grey' )
}

