% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot_functions.R
\name{blankRTplot}
\alias{blankRTplot}
\title{Blank response time plot}
\usage{
blankRTplot(tDim = c(0, 2), pDim = c(0, 1), ver = "CDF", unit = "s",
  ...)
}
\arguments{
\item{tDim}{the minimum and maximum for the time axis.}

\item{pDim}{the minimum and maximum for the choice/accuracy/probability
axis.}

\item{ver}{the type of plot to draw (i.e. 'CDF', 'PDF',
'HF', 'CAF', 'PvT', or 'blank' ).}

\item{unit}{the unit for the response times (e.g. 'ms' or 's').}

\item{...}{additional plotting parameters.}
}
\value{
A blank response time plot.
}
\description{
Creates a blank response time plot with standard labels.
}
\examples{
layout( rbind( 1:3, 4:6 ) )
blankRTplot(); blankRTplot(ver='PDF');
blankRTplot(ver='CAF'); blankRTplot(ver='HF');
blankRTplot(ver='PvT'); blankRTplot(ver='blank')
}

