% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot_functions.R
\name{CAF_points}
\alias{CAF_points}
\title{Conditional accuracy function}
\usage{
CAF_points(rt, ch, prb = seq(0.1, 0.9, 0.2), grp = NULL, plt = NULL,
  opt = c(T, F))
}
\arguments{
\item{rt}{vector of response times.}

\item{ch}{a vector of binary choices (i.e. 0 or 1).}

\item{prb}{the sequence of cumulative probabilities for which the
quantiles should be determined.}

\item{plt}{a named list that allows specification of graphical
parameters. The named list plt$ln can be used to set the graphical
elements (e.g.. the options \code{lwd} and \code{lty}) for the
drawn line. The named list plt$pt can be used to
set the graphical elements (e.g. the option \code{pch}) for the
drawn points.}

\item{opt}{logical vector; indicates if 1) the line should be drawn,
and 2) if output should be returned.}
}
\value{
A matrix giving the quantile estimates and the corresponding
  cumulative probabilities.
}
\description{
Draws the conditional accuracy at different response time quantiles.
}
\examples{
# Load in example dataset
data("priming_data")
d = priming_data
layout( cbind(1,2) )
# Single subject
sel = d$Condition == 2 & d$Subject == 6
rt = d$RT[sel]; ch = d$Accuracy[sel]
blankRTplot( xDim = c(.2,1), yDim=c(.5,1), ver='CAF')
plt = list( pt = list( pch = 19 ) )
CAF_points( rt, ch, plt = plt )
# Aggregating over multiple subjects
sel = d$Condition == 2
rt = d$RT[sel]; ch = d$Accuracy[sel]; grp = d$Subject[sel]
blankRTplot( xDim = c(.2,1), yDim = c(.5,1),
  bty = 'l', ver = 'QPE', cex.axis = 1.5,
  cex.lab = 1.5 )
plt = list( pt = list( bg = 'grey', pch = 23, cex = 1.2 ) )
CAF_points( rt, ch, grp = grp, plt = plt )
}

