% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot_functions.R
\name{cdf_curve}
\alias{cdf_curve}
\title{Joint CDF curves for response time and choice data.}
\usage{
cdf_curve(rt, ch, sel = 1, grp = NULL, plt = NULL, opt = c(T, T, F))
}
\arguments{
\item{rt}{vector of response times.}

\item{ch}{a vector of binary choices (i.e. 0 or 1).}

\item{sel}{If 0, the CDF for responses times when choice = 0 is
drawn. If 1, the CDF for responses times when choice = 1 is drawn.}

\item{grp}{an optional vector with a grouping factor (e.g. subjects).}

\item{plt}{a named list that allows specification of graphical
parameters. The named list plt$ln can be used to set the graphical
elements (e.g.. the options \code{lwd} and \code{lty}) for the
drawn line. The named lists plt$pt1 and plt$pt2 can be used to
set the graphical elements (e.g. the option \code{pch}) for the
points denoting the median and mean.}

\item{opt}{logical vector; indicates if 1) the joint distribution
should be used, 2) the line should be drawn, and 3) if output
should be returned.}
}
\value{
A list consisting of...
\describe{
  \item{\code{CDF}}{a matrix with the estimated cumulative
    probabilities and the corresonding resopnse times.}
  \item{\code{Median}}{the response time and cumulative
    probability for the median}
  \item{\code{Mean}}{the response time and cumulative
    probability for the mean}
  }
}
\description{
Draws a line for the estimated joint CDF of a set response
times and choices on an already existing plot.
}
\examples{
# Load in example dataset
data("priming_data")
d = priming_data
layout( cbind(1,2) )
# Single subject
sel = d$Condition == 4 & d$Subject == 1
rt = d$RT[sel]; ch = d$Accuracy[sel]
blankRTplot()
cdf_curve( rt, ch )
plt = list( ln = list( lty = 2 ) )
cdf_curve( rt, ch, sel = 0, plt = plt )
# Aggregating over multiple subjects
sel = d$Condition == 4
rt = d$RT[sel]; ch = d$Choice[sel]; grp = d$Subject[sel]
blankRTplot( bty = 'l', ver = 'CDF', cex.axis = 1.5, cex.lab = 1.5 )
plt = list( ln = list( lwd = 2 ), pt1 = list( bg = 'grey' ) )
cdf_curve( rt, ch, grp = grp, plt = plt )
plt = list( ln = list( lty = 2, lwd = 2 ), pt1 = list( bg = 'grey' ) )
cdf_curve( rt, ch, sel = 0, grp = grp, plt = plt )
}

