% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{draw_ui}
\alias{draw_ui}
\title{Draw Uncertainty Intervals}
\usage{
draw_ui(ui, val = NULL, ind = 1, flip = F, type = "segments", ...)
}
\arguments{
\item{ui}{a 'rtplots_ui' object.}

\item{val}{the choice/accuracy value for which to
draw the uncertainty intervals.}

\item{ind}{the index of the desired choice/accuracy
value over which to draw uncertainty intervals.}

\item{flip}{logical; if \code{true} the y-axis
values are flipped to be negative.}

\item{type}{the type of method used to draw the
uncertainty interval. Options are
\itemize{
  \item \code{segments}.
  \item \code{arrows}.
  \item \code{polygon}.
}}

\item{...}{additional plotting parameters. See
\code{\link[graphics]{segments}},
\code{\link[graphics]{arrows}}, and
\code{\link[graphics]{polygon}}.}
}
\description{
A convenience function that draws uncertainty
intervals computed from a 'rtplots' object onto
an already existing plot.
}
