% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot_functions.R
\name{hazard_curve}
\alias{hazard_curve}
\title{Estimated hazard function for response time and choice data.}
\usage{
hazard_curve(rt, ch, sel = 1, opt = c(T, T, F), ...)
}
\arguments{
\item{rt}{vector of response times.}

\item{ch}{a vector of binary choices (i.e. 0 or 1).}

\item{sel}{If 0, the PDF for responses times when choice = 0 is
drawn. If 1, the PDF for responses times when choice = 1 is drawn.}

\item{opt}{logical vector; indicates if 1) the joint distribution
should be used, 2) the line should be drawn, and 3) if output
should be returned.}

\item{...}{additional plotting parameters.}
}
\value{
A matrix giving the selected response times, density
  estimates, cumulative probabilities, and the estimated hazard
  function.
}
\description{
Draws a line for the estimated joint PDF of a set response
times and choices on an already existing plot.
}

