% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot_functions.R
\name{pdf_curve}
\alias{pdf_curve}
\title{PDF curves for response time and choice data}
\usage{
pdf_curve(rt, ch, sel = 1, grp = NULL, opt = list(), ...)
}
\arguments{
\item{rt}{vector of response times.}

\item{ch}{a vector of binary choices (i.e. 0 or 1).}

\item{sel}{If 0, the PDF for responses times when choice = 0 is
drawn. If 1, the PDF for responses times when choice = 1 is drawn.}

\item{grp}{an optional vector with a grouping factor (e.g. subjects).}

\item{opt}{a list of named options:
\describe{
  \item{\code{jnt}}{If true, the joint distribution is used.}
  \item{\code{draw}}{If true, the curve is drawn.}
  \item{\code{out}}{If true, output is returned.}
  \item{\code{flip}}{If true, the curve is flipped about the
    x-axis.}
}}

\item{...}{additional plotting parameters.}
}
\value{
A list consisting of...
\describe{
  \item{\code{pv}}{a data frame with the plotting values
    used for the x-axis and the y-axis.}
  \item{\code{g}}{when a grouping factor is present, a list
    with the matrix of y-axis values per level, the vector
    of associated response times, and the total number of
    observations per each included level}
  \item{\code{v}}{a list of additional variables, the total
    number of observations for all levels, the choice proportion
    for each level, the choice selection, and when there
    is no grouping factor, the response times and their
    associated estimated density.}
  \item{\code{i}}{a list of the input variables.}
  \item{\code{opt}}{a list of the options used.}
  }
}
\description{
Draws a line for the estimated PDF of a set response
times (conditioned on choice) on an already existing plot.
}
\examples{
# Load in example dataset
data("priming_data")
d = priming_data
layout( cbind(1,2) )
# Single subject
sel = d$Condition == 4 & d$Subject == 1
rt = d$RT[sel]; ch = d$Accuracy[sel]
blankRTplot( pDim = c(0,4), ver='PDF' )
pdf_curve( rt, ch )
pdf_curve( rt, ch, sel = 0, lty = 2 )
# Aggregating over multiple subjects
sel = d$Condition == 0 # Not all subjects had responses for each choice
rt = d$RT[sel]; ch = d$Choice[sel]; grp = d$Subject[sel]
blankRTplot( bty = 'l', ver = 'PDF', pDim = c(0,4), cex.axis = 1.5, cex.lab = 1.5 )
pdf_curve( rt, ch, grp = grp, lwd = 2 )
pdf_curve( rt, ch, sel = 0, grp = grp, lwd = 2, lty = 2 )
}

