% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot.rtplots}
\alias{plot.rtplots}
\title{Plot Method for 'rtplots' Object}
\usage{
\method{plot}{rtplots}(x, y = NULL, inc = NULL, type = NULL,
  xlab = NULL, ylab = NULL, bty = NULL, ...)
}
\arguments{
\item{x}{a 'rtplots' object.}

\item{y}{the y-axis boundaries.}

\item{inc}{the interval for the ticks on the
x and y-axes (depending on the plot).}

\item{type}{if eqaul to \code{'blank'}, a completely
blank plot is generated. Otherwise, converted to \code{'n'}.}

\item{xlab}{a character string for the label on the x-axis.}

\item{ylab}{a character string for the label on the y-axis.}

\item{bty}{the type of box to draw around the plot. See
\code{\link[graphics]{par}}.}

\item{...}{additional plotting parameters. See
\code{\link[graphics]{par}}.}
}
\description{
Creates a blank plot based on a 'rtplots' object.
}

