% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{pvt_plot}
\alias{pvt_plot}
\title{Create a Probability versus Time Figure}
\usage{
pvt_plot(object, keep = NULL, ind = NULL, val = NULL, type = "p",
  pch = NULL, cex = NULL, col = NULL, bg = NULL, lwd = NULL)
}
\arguments{
\item{object}{a rtplots object.}
}
\description{
A convenience function that draws points, lines, or
line segments for a probability versus time figure
onto an already existing plot. Allows fine-tune
control of plotting parameters such as point type,
size, etc.
}
