% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot_functions.R
\name{pvt_points}
\alias{pvt_points}
\title{Add points to a P(Y=1) by T(x) plot}
\usage{
pvt_points(rt, ch, cvrts, grp = NULL, T_x = mean, plt = list(),
  opt = list(), ...)
}
\arguments{
\item{rt}{vector of response times.}

\item{ch}{a vector of binary choices (i.e. 0 or 1).}

\item{grp}{an optional vector with a grouping factor (e.g. subjects).}

\item{T_x}{a function to calculate a text statistic over
a vector (e.g. mean, median, etc.).}

\item{plt}{an optional list giving values for the \code{pch},
\code{col}, and \code{bg} plotting parameters.}

\item{opt}{a list of named options:
\describe{
  \item{\code{draw}}{If true, the intervals are drawn.}
  \item{\code{out}}{If true, output is returned.}
}}

\item{...}{additional plotting parameters.}
}
\value{
A list consisting of ...
\describe{
  \item{\code{pv}}{a list with the plotting values used for
    the x-axis and the y-axis.}
  \item{\code{g}}{when a grouping factor is present, a list
    with the data-frame for the y-axis values and the x-axis
    values by group and covariate level.}
  \item{\code{i}}{a list of the input variables.}
  \item{\code{opt}}{a list of the options used.}
  }
}
\description{
Adds a set of points for a test statistic for response times
calculated over the levels of a covariate against the
probability of picking a '1' response.
}

