% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot_functions.R
\name{quantile_points}
\alias{quantile_points}
\title{Quantile-probability estimates}
\usage{
quantile_points(rt, ch, sel = 1, prb = seq(0.1, 0.9, 0.2), grp = NULL,
  plt = NULL, opt = c(T, T, F))
}
\arguments{
\item{rt}{vector of response times.}

\item{ch}{a vector of binary choices (i.e. 0 or 1).}

\item{sel}{If 0, the quantile-probability estimates for responses
times when choice = 0 are drawn. If 1, the quantile-probability
estimates for responses times when choice = 1 are drawn.}

\item{prb}{the sequence of cumulative probabilities for which the
quantiles should be determined.}

\item{plt}{a named list that allows specification of graphical
parameters. The named list plt$ln can be used to set the graphical
elements (e.g.. the options \code{lwd} and \code{lty}) for the
drawn line. The named list plt$pt can be used to
set the graphical elements (e.g. the option \code{pch}) for the
drawn points.}

\item{opt}{logical vector; indicates if 1) the joint distribution
should be used, 2) the line should be drawn, and 3) if output
should be returned.}
}
\value{
A matrix giving the quantile estimates and the corresponding
  cumulative probabilities.
}
\description{
Draws the estimated quantiles based on the corresponding joint
cumulative probabilities for a set of response times and choices.
}
\examples{
# Load in example dataset
data("priming_data")
d = priming_data
layout( cbind(1,2) )
# Single subject
sel = d$Condition == 6 & d$Subject == 20
rt = d$RT[sel]; ch = d$Accuracy[sel]
blankRTplot(xDim = c(0,1), ver='QPE')
quantile_points( rt, ch )
plt = list( ln = list( type = 'n' ), pt = list( pch = 18,
  col = 'red' ) )
quantile_points( rt, ch, sel = 0, plt = plt )
# Aggregating over multiple subjects
sel = d$Condition == 6
rt = d$RT[sel]; ch = d$Choice[sel]; grp = d$Subject[sel]
blankRTplot( xDim = c(0,1), bty = 'l', ver = 'QPE', cex.axis = 1.5,
  cex.lab = 1.5 )
plt = list( ln = list( type = 'n' ),
  pt = list( bg = 'grey', pch = 21, cex = 1.5 ) )
quantile_points( rt, ch, sel = 1, plt = plt )
plt = list( ln = list( type = 'n' ),
  pt = list( bg = 'white', pch = 22, cex = 1.5 ) )
quantile_points( rt, ch, sel = 0, plt = plt )
}

