% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{rtplots}
\alias{is.rtplots}
\alias{rtplots}
\title{Creates a 'rtplots' Object}
\usage{
rtplots(df, label = NULL, keep = NULL, type = "CDF", level = NULL, ...)

is.rtplots(x)
}
\arguments{
\item{df}{a data frame that includes (at a minimum) a column of
response times.}

\item{label}{a character string giving the column name for
the respons time variable, and optionally the choice/accuracy
variable and a grouping factor variable (e.g., subjects).
If \code{NULL}, the algorithm will attempt to locate response
time and choice/accuracy variables via a set of default names.}

\item{keep}{a logical vector, indicating which rows should be
kept when extracting the variables.}

\item{type}{the type of function to compute. Options are
\itemize{
  \item \code{CDF}; the cumulative distribution function.
  \item \code{QPE}; the quantile function.
}}

\item{level}{an optional index indicating the specific level
of the grouping factor to consider. This allows, for instance,
selecting a single subject.}

\item{...}{additional parameters for the density, distribution,
and quantile functions.}
}
\value{
An object of class 'rtplots', a list consisting of...
  \itemize{
    \item \code{pd}; a data frame with the x and y values
      to plot and their associated choice/accuracy value.
    \item \code{xm}; If relevant, a matrix whose columns
      contain the individual x-axis values of the function
      which were collapsed over the grouping factor.
    \item \code{ym}; If relevant, a matrix whose columns
      contain the individual y-axis values of the function
      which were collapsed over the grouping factor.
    \item \code{npfd}; a data frame with the number of
      observations, proportions, and frequencies based on
      the choice/accuracy values.
    \item \code{type}; the type of function that was applied.
    \item \code{x}; the range of x-axis values.
    \item \code{y}; the range of y-axis values.
  }
}
\description{
Creates a 'rtplots' object from a data frame of
response time and choice/accuracy data, allowing
for subsequent plotting.
}
\details{
Forthcoming
}
\examples{
# Load in example dataset
data("priming_data")
lbl = c( 'RT', 'Accuracy', 'Subjects' )
keep = priming_data$Condition == 0
obj = rtplots( priming_data, label = lbl, keep = keep )

}
\references{
Forthcoming
}

