% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{uncertainty}
\alias{uncertainty}
\title{Computes Uncertainty Intervals for 'rtplots' Objects}
\usage{
uncertainty(object, f = NULL, alpha = 0.95, ...)
}
\arguments{
\item{object}{a 'rtplots' object.}

\item{f}{a function to compute the boundaries of
the uncertainty interval given the desired
range of coverage (takes at least two parameters,
a vector \code{x} and a range \code{interval}).}

\item{alpha}{the desired width of coverate for the
given uncertainty interval.}

\item{...}{additional parameters for the
function \code{f}.}
}
\value{
Forthcoming
}
\description{
A function that computes a specified uncertainty
interval (such as \eqn{\pm} 2 standard errors from
the mean) for a 'rtplots' object.
}
