% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot_functions.R
\name{blankRTplot}
\alias{blankRTplot}
\title{Blank response time plot}
\usage{
blankRTplot(xDim = c(0, 2), yDim = c(0, 1), ver = "CDF", unit = "s",
  ...)
}
\arguments{
\item{xDim}{the minimum and maximum x-axis values.}

\item{yDim}{the minimum and maximum y-axis values.}

\item{ver}{the type of plot to draw (e.g. 'CDF', 'PDF',
'QPE', or 'CAF').}

\item{unit}{the unit for the response times (e.g. 'ms' or 's').}
}
\value{
A blank response time plot.
}
\description{
Creates a blank response time plot with standard labels.
}

