% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot_functions.R
\name{caf_points}
\alias{caf_points}
\title{Conditional accuracy function}
\usage{
caf_points(rt, ch, prb = seq(0.1, 0.9, 0.2), grp = NULL, opt = list(),
  ...)
}
\arguments{
\item{rt}{vector of response times.}

\item{ch}{a vector of binary choices (i.e. 0 or 1).}

\item{prb}{the sequence of cumulative probabilities for which the
quantiles should be determined.}

\item{grp}{an optional vector with a grouping factor (e.g. subjects).}

\item{opt}{a list of named options:
\describe{
  \item{\code{jnt}}{If true, the joint distribution is used.}
  \item{\code{draw}}{If true, the curve is drawn.}
  \item{\code{out}}{If true, output is returned.}
  \item{\code{flip}}{If true, the curve is flipped about the
    x-axis.}
  \item{\code{pts}}{If true, draw points instead of lines.}
}}

\item{...}{additional plotting parameters.}
}
\value{
A list consisting of...
\describe{
  \item{\code{pv}}{a data frame with the plotting values
    used for the x-axis and the y-axis.}
  \item{\code{g}}{when a grouping factor is present, a list
    with the matrix of y-axis values per level and the
    associated matrix of quantiles.}
  \item{\code{v}}{a list of additional variables, the choice
    proportion for each level and the choice selection.}
  \item{\code{i}}{a list of the input variables.}
  \item{\code{opt}}{a list of the options used.}
  }
}
\description{
Draws the conditional accuracy at different response time quantiles.
}
\examples{
# Load in example dataset
data("priming_data")
d = priming_data
layout( cbind(1,2) )
# Single subject
sel = d$Condition == 12 & d$Subject == 7
rt = d$RT[sel]; ch = d$Accuracy[sel]
blankRTplot( tDim = c(.6,1.6), ver = 'CAF' )
caf_points( rt, ch, opt = list( pts = F ) )
caf_points( rt, ch, pch = 21, bg = 'white' )
# Aggregating over multiple subjects
sel = d$Condition == 12
rt = d$RT[sel]; ch = d$Choice[sel]; grp = d$Subject[sel]
blankRTplot( tDim = c(0,1.2), ver = 'CAF', cex.axis = 1.5, cex.lab = 1.5 )
caf_points( rt, ch, grp = grp, opt = list( pts = F ), lwd = 2 )
caf_points( rt, ch, grp = grp, pch = 21, bg = 'white', cex = 1.2 )
}

