% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_plot_functions.R
\name{pdf_curve}
\alias{pdf_curve}
\title{Joint PDF curves for response time and choice data.}
\usage{
pdf_curve(rt, ch, sel = 1, grp = NULL, plt = NULL, opt = c(T, T, F))
}
\arguments{
\item{rt}{vector of response times.}

\item{ch}{a vector of binary choices (i.e. 0 or 1).}

\item{sel}{If 0, the PDF for responses times when choice = 0 is
drawn. If 1, the PDF for responses times when choice = 1 is drawn.}

\item{grp}{an optional vector with a grouping factor (e.g. subjects).}

\item{plt}{a named list that allows specification of graphical
parameters. The named list plt$ln can be used to set the graphical
elements (e.g.. the options \code{lwd} and \code{lty}) for the
drawn line. The named lists plt$pt1 and plt$pt2 can be used to
set the graphical elements (e.g. the option \code{pch}) for the
points denoting the median and mean.}

\item{opt}{logical vector; indicates if 1) the joint distribution
should be used, 2) the line should be drawn, and 3) if output
should be returned.}
}
\value{
A list consisting of...
\describe{
  \item{\code{PDF}}{a matrix with the estimated density}
  \item{\code{Mode}}{the response time and likelihood for the mode}
  \item{\code{Mean}}{the response time and likelihood for the mean}
  }
}
\description{
Draws a line for the estimated joint PDF of a set response
times and choices on an already existing plot.
}
\examples{
# Load in example data
data(priming_data)
d = priming_data
layout( cbind(1,2) )
# Single subject
sel = d$Condition == 5 & d$Subject == 16
rt = d$RT[sel]; ch = d$Accuracy[sel]
blankRTplot(yDim=c(0,3),ver='PDF')
pdf_curve( rt, ch )
plt = list( ln = list( col = 'blue' ) )
pdf_curve( rt, ch, sel = 0, plt = plt )
# Aggregating over multiple subjects
sel = d$Condition == 5
rt = d$RT[sel]; ch = d$Choice[sel]; grp = d$Subject[sel]
blankRTplot( yDim = c(0,2), bty = 'l', ver = 'CDF', cex.axis = 1.5,
  cex.lab = 1.5 )
plt = list( ln = list( lwd = 2 ), pt1 = list( bg = 'grey', cex = 1.5 ),
  pt2 = list( type = 'n' ) )
pdf_curve( rt, ch, grp = grp, plt = plt )
plt = list( ln = list( lty = 2, lwd = 2 ),
  pt1 = list( pch = 15, cex = 1.5 ),
  pt2 = list( type = 'n' ) )
pdf_curve( rt, ch, sel = 0, grp = grp, plt = plt )
}

